requirements.txt
openai-agents>=0.0.17
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0